Rename a/js/entry_points -> a/js/vue_instances to make it more obvious
