+++ app/javascript/vue_instances/ask.js
