+++ app/javascript/vue_instances/browse.js
