+++ app/javascript/vue_instances/fileUpload.js
