+++ app/javascript/vue_instances/index.js
