+++ app/javascript/vue_instances/navBar.js
