+++ app/javascript/vue_instances/notice.js
