+++ app/javascript/vue_instances/offer.js
