+++ app/views/asks/new.html.erb
@@ -21,7 +21,7 @@
 
 <script>
   document.addEventListener('DOMContentLoaded', () => {
-    EntryPoints.ask('#ask', <%= raw @json %>)
+    VueInstances.ask('#ask', <%= raw @json %>)
   })
 </script>
