+++ app/views/offers/new.html.erb
@@ -21,7 +21,7 @@
 
 <script>
   document.addEventListener('DOMContentLoaded', () => {
-    EntryPoints.offer('#offer', <%= raw @json %>)
+    VueInstances.offer('#offer', <%= raw @json %>)
   })
 </script>
