+++ app/views/public/landing_page.html.erb
@@ -2,6 +2,6 @@
 
 <script>
   document.addEventListener('DOMContentLoaded', () => {
-    EntryPoints.orientation('#orientation', <%= raw @json %>)
+    VueInstances.orientation('#orientation', <%= raw @json %>)
   })
 </script>
