+++ app/views/submission_response_imports/new.html.erb
@@ -9,7 +9,7 @@
 
 <script>
   document.addEventListener('DOMContentLoaded', () => {
-    EntryPoints.fileUpload('#upload', {
+    VueInstances.fileUpload('#upload', {
       fieldName: 'submission_response_import[file]',
     })
   })
