+++ doc/architecture.md
@@ -19,6 +19,6 @@ We generally use rails norms and idioms, on top of which we've introduced a few 
 * Nested resource composition and hydration using a [simplified Form](https://github.com/rubyforgood/mutual-aid/commit/6741565ffc265689f859e77932747f53e4c03051#diff-b29d1eb8e34d510b96985edf0d9211a2ad4f39df24e74adcbdc66c769db68547). This is a little bit like the FormObject pattern but with fewer responsibilities. Of the various uses [listed here](https://selleo.com/blog/essential-rubyonrails-patterns-form-objects), our implementation only addresses _Form Object as virtual resource_ and _Form Object as composed resource_. Note that we're repurposing the aforementioned ActiveInteraction gem in our forms, mostly because it provides some convenient coercions.
 
 ## Other architectural notes
-* All our pages are server-rendered, but some include Vue instances. Some of these 'apps' take over small sections of the page (such as the NavBar) while others take over most of the page. The entry point into all Vue instances can be found in [this directory](https://github.com/rubyforgood/mutual-aid/tree/main/app/javascript/entry_points), and here's [one example](https://github.com/rubyforgood/mutual-aid/blob/main/app/views/asks/new.html.erb#L24) of how an instance is included on an ERB template.
+* All our pages are server-rendered, but some include Vue instances. Some of these 'apps' take over small sections of the page (such as the NavBar) while others take over most of the page. The entry point into all Vue instances can be found in [this directory](https://github.com/rubyforgood/mutual-aid/tree/main/app/javascript/vue_instances), and here's [one example](https://github.com/rubyforgood/mutual-aid/blob/main/app/views/asks/new.html.erb#L24) of how an instance is included on an ERB template.
 * Our [Contributions](http://mutual-aid-demo.herokuapp.com/contributions) page supports some complex filtering, the design of which is [described here](doc/structure-of-contribution-filters.md).
